Adds auto-saving in Pong.
The state of a Pong game is now saved whenever the onPause method is
entered.  Currently the old AGM comma- and colon- separated strings are
used to store the information.

There may be some unintentional effect of this that causes lag for users
when reloading the game; see: https://github.com/CovenantCS/AGM/issues/3