Bowling implementation
Start of work implementing Bowling game.  Currently displays, but does
not scale properly to screen size.  Also displays no response to user
input.

Also modified sprites that contain a color value to display properly
(required adding an alpha channel value to the packed ints) and set the
onDraw in AndroidBoardViewer to use the sprite's color-value instead of
making everything Red.  Leaving it in this branch because I haven't
thoroughly tested it; not sure I got everything.